lazy: share the direct-read machinery, support gemma4 PLE tables
Move the row reader into arch-agnostic src/llama-lazy-reader.h, owned
by llama_model_base keyed by tensor name; arches opt in with one
load_lazy_reader call. qwen4exp now uses the shared reader, and
gemma4 stages its per-layer rows the same way. Gemma's ~5 KiB
contiguous rows show no prefill win either way, which is why the
reader stays opt-in per tensor rather than unconditional.

Also fix the missing <string> include in llama-mmap.h (MSVC needs it
for the name() declaration) and make the help text generic so it does
not stale when arches opt in.
