+++ src/models/qwen4exp.cpp
@@ -4,155 +4,7 @@
 #include "llama-memory-recurrent.h"
 
 #include <algorithm>
-#include <cerrno>
 #include <cinttypes>
-#include <cstring>
-#include <exception>
-#include <stdexcept>
-#include <thread>
-#include <utility>
-#include <vector>
-
-
-#ifndef _WIN32
-#include <fcntl.h>
-#include <unistd.h>
-#endif
-
-llama_model_qwen4exp::llama_model_qwen4exp(const struct llama_model_params & params) : llama_model_base(params) {}
-llama_model_qwen4exp::~llama_model_qwen4exp() = default;
-
-#ifndef _WIN32
-// Direct-read path for the lazy PLE table (--lazy-mode on-direct).
-// The n-gram row indices of a whole ubatch are known host-side before the graph
-// runs, so the rows can be read with explicit pread()s of exactly what the gather
-// needs instead of demand-faulting file pages in through the mmap. Rows are sorted
-// (dedup + ascending file offsets) and read by a small set of worker threads, so
-// the block layer sees a sorted, parallel batch instead of one serialized page
-// fault per ~90-byte row. The table stays on disk; the kernel page cache provides
-// reuse across ubatches.
-struct llama_model_qwen4exp::ple_direct_reader {
-    ple_direct_reader(int fd, size_t base, size_t row_size, int64_t n_rows, int n_threads,
-                      enum ggml_type type, int64_t head_dim)
-        : fd(fd), base(base), row_size(row_size), n_rows(n_rows), n_threads(n_threads),
-          head_dim(head_dim), to_float(type == GGML_TYPE_F32 ? nullptr : ggml_get_type_traits(type)->to_float) {
-        // F32 rows have no dequantizer; they are staged as-is, like ggml_get_rows
-        GGML_ASSERT((type == GGML_TYPE_F32 || to_float != nullptr) && head_dim > 0);
-    }
-    ~ple_direct_reader() {
-        if (fd >= 0) {
-            ::close(fd);
-        }
-    }
-
-    const int      fd;
-    const size_t   base;       // file offset of row 0
-    const size_t   row_size;   // bytes per quantized row
-    const int64_t  n_rows;
-    const int      n_threads;  // in-flight read workers
-    const int64_t  head_dim;   // F32 elements per staged row
-    ggml_to_float_t to_float;  // same dequantizer the ggml_get_rows CPU kernel uses
-
-    // fill dst with the n gathered rows, dequantized to F32:
-    // dst[slot * head_dim, ...) = to_float(table[rows[slot]])
-    // throws on IO errors; never lets an exception escape a worker thread
-    void gather(const int32_t * rows, int64_t n, float * dst) const {
-        std::vector<std::pair<int32_t, int32_t>> pairs; // (row, dst slot)
-        pairs.reserve(n);
-        for (int64_t i = 0; i < n; ++i) {
-            GGML_ASSERT(rows[i] >= 0 && (int64_t) rows[i] < n_rows);
-            pairs.emplace_back(rows[i], (int32_t) i);
-        }
-
-        std::sort(pairs.begin(), pairs.end()); // equal rows adjacent, file order
-
-        // decodes gather a handful of rows; threads are not worth it there
-        const int n_workers = (int) std::min<int64_t>(n_threads, std::max<int64_t>(1, n / 32));
-
-        // worker w reads rows pairs[n*w/n_workers, n*(w+1)/n_workers)
-        auto run_chunk = [&](int w, std::exception_ptr & err) {
-            try {
-                run_range(pairs, n * w / n_workers, n * (w + 1) / n_workers, dst);
-            } catch (...) {
-                err = std::current_exception();
-            }
-        };
-
-        // an exception leaving a joinable std::thread, or destroying one,
-        // terminates the process; keep worker creation failure-safe
-        std::vector<std::exception_ptr> errs(n_workers);
-        std::vector<std::thread> workers;
-        try {
-            for (int w = 1; w < n_workers; ++w) {
-                workers.emplace_back([&run_chunk, &errs, w]() {
-                    run_chunk(w, errs[w]);
-                });
-            }
-        } catch (...) {
-            for (auto & t : workers) {
-                t.join();
-            }
-            throw;
-        }
-
-        run_chunk(0, errs[0]); // this thread takes the first chunk
-        for (auto & t : workers) {
-            t.join();
-        }
-
-        for (const auto & err : errs) {
-            if (err) {
-                std::rethrow_exception(err);
-            }
-        }
-    }
-
-private:
-    void run_range(const std::vector<std::pair<int32_t, int32_t>> & pairs,
-                   int64_t begin, int64_t end, float * dst) const {
-        std::vector<uint8_t> bounce(row_size);
-        for (int64_t i = begin; i < end; ) {
-            int64_t j = i;
-            while (j + 1 < end && pairs[j + 1].first == pairs[i].first) {
-                ++j; // dedup: one read serves the whole run
-            }
-            const size_t off = base + (size_t) pairs[i].first * row_size;
-            for (size_t done = 0; done < row_size; ) {
-                const ssize_t n_read = ::pread(fd, bounce.data() + done, row_size - done, off + done);
-                if (n_read < 0 && errno == EINTR) {
-                    continue; // interrupted by a signal without SA_RESTART
-                }
-                if (n_read <= 0) {
-                    throw std::runtime_error(format("PLE direct read of %zu bytes at file offset %zu failed: %s",
-                            row_size, off, n_read == 0 ? "unexpected EOF" : strerror(errno)));
-                }
-                done += n_read;
-            }
-            float * first = dst + (size_t) pairs[i].second * head_dim;
-            if (to_float) {
-                to_float(bounce.data(), first, head_dim);
-            } else {
-                memcpy(first, bounce.data(), (size_t) head_dim * sizeof(float));
-            }
-            for (int64_t k = i + 1; k <= j; ++k) {
-                memcpy(dst + (size_t) pairs[k].second * head_dim, first, (size_t) head_dim * sizeof(float));
-            }
-            i = j + 1;
-        }
-    }
-};
-
-// matching interface so the call sites compile; never constructed on this
-// platform, see load_arch_tensors
-#else
-struct llama_model_qwen4exp::ple_direct_reader {
-    const int64_t head_dim = 0;
-
-    void gather(const int32_t *, int64_t, float *) const {
-        GGML_ABORT("PLE direct reads are not supported on this platform");
-    }
-};
-#endif
 
 // bad metadata must be catchable: GGML_ASSERT aborts the whole process
 static void qwen4exp_require_nonzero(const llama_model_loader & ml, llm_kv kid, uint32_t value) {
@@ -334,43 +186,10 @@ void llama_model_qwen4exp::load_arch_tensors(llama_model_loader & ml) {
 
         per_layer_tok_embd = create_tensor(tn(LLM_TENSOR_PER_LAYER_TOKEN_EMBD, "weight"),
                                            { hparams.ple_head_dim, ple_rows }, TENSOR_READ_LAZY);
+
         // --lazy-mode on-direct: read the gathered rows with explicit pread()s
         // instead of faulting them in through the mmap
-        if (ml.lazy.mode == LLAMA_LAZY_MODE_DIRECT) {
-#ifndef _WIN32
-            const auto & ple_w = ml.require_weight(ple_name.c_str());
-            // an independent buffered descriptor: dup() would share the loader's
-            // open file description, whose readahead advice and O_DIRECT flag
-            // (init_mappings applies POSIX_FADV_SEQUENTIAL, --load-mode dio)
-            // would fight the small scattered row reads
-            const int fd = ::open(ml.files[ple_w.idx]->name().c_str(), O_RDONLY | O_CLOEXEC);
-            if (fd < 0) {
-                // e.g. a FILE*-backed model has no reopenable path; the tensor is
-                // still lazy, so keep serving it through the mmap reads
-                LLAMA_LOG_WARN("%s: could not open %s for direct reads (%s), using lazy mmap reads\n",
-                        __func__, ml.files[ple_w.idx]->name().c_str(), strerror(errno));
-            } else {
-#ifdef __linux__
-                // rows are tiny and scattered, so sequential readahead would be
-                // pure waste; Darwin has no posix_fadvise
-                ::posix_fadvise(fd, 0, 0, POSIX_FADV_RANDOM);
-#endif
-
-                // in-flight reads are IO queue depth, not compute; 2x cores worked
-                // well on NVMe and stays sane on smaller machines
-                const int n_threads = 2 * (int) std::max(1u, std::thread::hardware_concurrency());
-
-                ple_reader = std::make_unique<ple_direct_reader>(fd, ple_w.offs,
-                        ggml_row_size(per_layer_tok_embd->type, per_layer_tok_embd->ne[0]), ple_rows, n_threads,
-                        per_layer_tok_embd->type, hparams.ple_head_dim);
-
-                LLAMA_LOG_INFO("%s: PLE direct read enabled: %" PRId64 " rows of %zu bytes at file offset %zu, %d threads\n",
-                        __func__, ple_rows, ple_reader->row_size, ple_w.offs, n_threads);
-            }
-#else
-            LLAMA_LOG_WARN("%s: --lazy-mode on-direct is not supported on this platform, using lazy mmap reads\n", __func__);
-#endif
-        }
+        ple_reader = load_lazy_reader(ml, ple_name.c_str(), per_layer_tok_embd);
     }
 
     for (int il = 0; il < n_layer; ++il) {
@@ -1370,7 +1189,7 @@ ggml_tensor * llama_model_qwen4exp::graph::build_inp_ple(
 
     if (static_cast<const llama_model_qwen4exp &>(model).ple_reader) {
         // direct-read mode: set_input() pre-gathers the rows host-side, so the
-        // staged tensor replaces ggml_get_rows and the table pages stay untouched
+        // staged tensor replaces ggml_get_rows and the table pages stay untouched.
         // F32 matches the ggml_get_rows output type, so the downstream mul_mats
         // take the same kernels as the baseline path
         ple_inp->data = ggml_new_tensor_2d(ctx0, GGML_TYPE_F32,
